ai-service/pyproject.toml
[project]
name = "wsc3d-ai-service"
version = "0.1.0"
description = "FastAPI AI helpers for WSC3D annotation"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "numpy==1.26.4",
  "pillow",
  "opencv-python==4.10.0.84",
  "torch",
  "torchvision",
  "timm",
  "ultralytics",
  "sam3 @ git+https://github.com/facebookresearch/sam3.git",
  "ftfy==6.1.1",
  "triton-windows; platform_system == 'Windows'",
  "mobile-sam @ git+https://github.com/ChaoningZhang/MobileSAM.git"
]

[tool.uv]
package = false
